insn: fix I- versus R-type ADD/ADDI mess up
I accidentally confused the I-type and R-type arithmetic instructions,
defining constants for the R-type ones, but using them to parse the
I-type instructions instead. This was unnoticed because the FUNC3s are
actually the same for ADDI and ADD. Oops.

Since they'll all need to be implemented anyway, just add the other
insn's parsing routines anyway.

Signed-off-by: Conor Dooley <mail@conchuod.ie>