treewide: abstract memory/register implementation away from insn handling
Due to MMIO peripherals, it's not going to be possible to implement
"memory" as an array, so ditch that idea from the start. While we are at
it, abstract away the register interface too, so that it can be
implemented in a more idiomatic way in the future without disturbing the
instruction parsing code too much.

Signed-off-by: Conor Dooley <mail@conchuod.ie>