treewide: implement a bus type & make memory accesses use it
Some sort of generic memory & memory-mapped I/O is clearly going to be
needed for the various parts of this project.
Implement it using generic const expressions, because that is fuin and
something that I have yet to use in Rust! Also it looks oh-so-much
neater than having to implement a read & write for every type.

There are two new modules that come with this:
- bus, so that importing the bus is independent of any other stuff, as
  it should be
- lebytes, because there is no trait that covers T::{to,from}_le_bytes()
  so I have to provide it myself

Signed-off-by: Conor Dooley <mail@conchuod.ie>