rustfmt: drop line length to 80 chars
There's a few other things in here, although mostly they're just me
adding the default.
The main changes are the drop to 80 character line lengths, but also
removing the relative line-length stuff that rust derives from the max,
so that it is now 80 for everything.
`struct_lit_single_line = false` is technically new, but all the
initialisers currently in the code do this anyway, although rustfmt
could have done something different with them. Now it can't.

Signed-off-by: Conor Dooley <mail@conchuod.ie>